+++ bsp/ti-tms320c6678/applications/board.c
@@ -9,27 +9,11 @@
  */
 
 #include "board.h"
-#include "trap.h"
 #include "interrupt.h"
 #include "drv_timer.h"
 #include "KeyStone_common.h"
 
 #include <rtthread.h>
-
-/**
- * This is the timer interrupt service routine.
- *
- */
-void rt_hw_systick_isr(void)
-{
-	/* enter interrupt */
-	rt_interrupt_enter();
-
-	rt_tick_increase();
-
-	/* leave interrupt */
-	rt_interrupt_leave();
-}
 
 /**
  * This function will initial board.
@@ -42,16 +26,12 @@ void rt_hw_board_init(void)
 	// initial interrupt controller
 	rt_hw_interrupt_init();
 
-	// initial system trap
-	rt_trap_init();
-
 	// initial system timer
-	hw_system_timer_init();
+    hw_system_timer_init();
 
-	/* initialize memory system */
-	rt_kprintf("heap: 0x%08x - 0x%08x\n", RT_HW_HEAP_BEGIN, RT_HW_HEAP_END);
-	rt_system_heap_init(RT_HW_HEAP_BEGIN, RT_HW_HEAP_END);
+    /* initialize memory system */
+    rt_kprintf("heap: 0x%08x - 0x%08x\n", RT_HW_HEAP_BEGIN, RT_HW_HEAP_END);
+    rt_system_heap_init(RT_HW_HEAP_BEGIN, RT_HW_HEAP_END);
 
-	hw_system_timer_start();
+    hw_system_timer_start();
 }
-
