+++ libcpu/ti-dsp/c6x/context.asm
@@ -9,7 +9,7 @@
 ;
 
 ;-----------------------------------------------------------
-; context switch for C6678 DSP
+; 				context switch for C6000 DSP
 ;-----------------------------------------------------------
 
 	.include "contextinc.asm"
