+++ libcpu/ti-dsp/c6x/intexc.asm
@@ -9,7 +9,7 @@
 ;
 
 ;-----------------------------------------------------------
-; interrupt and execption handler for C6678 DSP
+; 		interrupt and execption handler for C000 DSP
 ;-----------------------------------------------------------
 
 ;-----------------------------------------------------------
