+++ libcpu/ti-dsp/c6x/vector.asm
@@ -9,7 +9,7 @@
 ;
 
 ;-----------------------------------------------------------
-; interrupt vector table for C6678 DSP
+; 			interrupt vector table for C6000 DSP
 ;-----------------------------------------------------------
 
 ;-----------------------------------------------------------
